+++ presto-main/src/main/java/com/facebook/presto/memory/LocalMemoryManager.java
@@ -38,22 +38,15 @@ public final class LocalMemoryManager
 {
     public static final MemoryPoolId GENERAL_POOL = new MemoryPoolId("general");
     public static final MemoryPoolId RESERVED_POOL = new MemoryPoolId("reserved");
-    public static final MemoryPoolId SYSTEM_POOL = new MemoryPoolId("system");
 
     private DataSize maxMemory;
     private Map<MemoryPoolId, MemoryPool> pools;
 
     @Inject
-    public LocalMemoryManager(NodeMemoryConfig config, ReservedSystemMemoryConfig systemMemoryConfig)
+    public LocalMemoryManager(NodeMemoryConfig config)
     {
         requireNonNull(config, "config is null");
-        long availableMemory = Runtime.getRuntime().maxMemory();
-        if (config.isLegacySystemPoolEnabled()) {
-            configureLegacyMemoryPools(config, systemMemoryConfig, availableMemory);
-        }
-        else {
-            configureMemoryPools(config, availableMemory);
-        }
+        configureMemoryPools(config, Runtime.getRuntime().maxMemory());
     }
 
     private void configureMemoryPools(NodeMemoryConfig config, long availableMemory)
@@ -73,23 +66,6 @@ public final class LocalMemoryManager
         }
         verify(generalPoolSize > 0, "general memory pool size is 0");
         builder.put(GENERAL_POOL, new MemoryPool(GENERAL_POOL, new DataSize(generalPoolSize, BYTE)));
-        this.pools = builder.build();
-    }
-
-    private void configureLegacyMemoryPools(NodeMemoryConfig config, ReservedSystemMemoryConfig systemMemoryConfig, long availableMemory)
-    {
-        checkArgument(systemMemoryConfig.getReservedSystemMemory().toBytes() < availableMemory, "Reserved memory %s is greater than available heap %s", systemMemoryConfig.getReservedSystemMemory(), new DataSize(availableMemory, BYTE));
-        maxMemory = new DataSize(availableMemory - systemMemoryConfig.getReservedSystemMemory().toBytes(), BYTE);
-
-        ImmutableMap.Builder<MemoryPoolId, MemoryPool> builder = ImmutableMap.builder();
-        checkArgument(config.getMaxQueryMemoryPerNode().toBytes() <= maxMemory.toBytes(), format("%s set to %s, but only %s of useable heap available", QUERY_MAX_MEMORY_PER_NODE_CONFIG, config.getMaxQueryMemoryPerNode(), maxMemory));
-        long generalPoolSize = maxMemory.toBytes();
-        if (config.isReservedPoolEnabled()) {
-            builder.put(RESERVED_POOL, new MemoryPool(RESERVED_POOL, config.getMaxQueryMemoryPerNode()));
-            generalPoolSize -= config.getMaxQueryMemoryPerNode().toBytes();
-        }
-        builder.put(GENERAL_POOL, new MemoryPool(GENERAL_POOL, new DataSize(generalPoolSize, BYTE)));
-        builder.put(SYSTEM_POOL, new MemoryPool(SYSTEM_POOL, systemMemoryConfig.getReservedSystemMemory()));
         this.pools = builder.build();
     }
 
@@ -126,11 +102,6 @@ public final class LocalMemoryManager
     public MemoryPool getGeneralPool()
     {
         return pools.get(GENERAL_POOL);
-    }
-
-    public Optional<MemoryPool> getSystemPool()
-    {
-        return Optional.ofNullable(pools.get(SYSTEM_POOL));
     }
 
     public Optional<MemoryPool> getReservedPool()
